+++ src/main/java/org/sonar/plugins/delphi/symbol/declaration/PropertyNameDeclaration.java
@@ -4,9 +4,14 @@ import com.google.common.collect.ComparisonChain;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+import net.sourceforge.pmd.lang.ast.Node;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import org.sonar.plugins.delphi.antlr.ast.node.NameReferenceNode;
+import org.sonar.plugins.delphi.antlr.ast.node.PrimaryExpressionNode;
 import org.sonar.plugins.delphi.antlr.ast.node.PropertyNode;
+import org.sonar.plugins.delphi.antlr.ast.node.PropertyReadSpecifierNode;
+import org.sonar.plugins.delphi.antlr.ast.node.PropertyWriteSpecifierNode;
 import org.sonar.plugins.delphi.symbol.SymbolicNode;
 import org.sonar.plugins.delphi.symbol.resolve.Invocable;
 import org.sonar.plugins.delphi.type.Type;
@@ -19,6 +24,8 @@ public final class PropertyNameDeclaration extends AbstractDelphiNameDeclaration
   private final boolean isClassInvocable;
   private final boolean isDefaultProperty;
   private final Type type;
+  private final DelphiNameDeclaration readDeclaration;
+  private final DelphiNameDeclaration writeDeclaration;
   private int hashCode;
 
   public PropertyNameDeclaration(
@@ -28,7 +35,9 @@ public final class PropertyNameDeclaration extends AbstractDelphiNameDeclaration
         extractParameters(node, concreteDeclaration),
         node.isClassProperty(),
         node.isDefaultProperty(),
-        extractType(node, concreteDeclaration));
+        extractType(node, concreteDeclaration),
+        extractReadDeclaration(node, concreteDeclaration),
+        extractWriteDeclaration(node, concreteDeclaration));
   }
 
   private PropertyNameDeclaration(
@@ -36,12 +45,16 @@ public final class PropertyNameDeclaration extends AbstractDelphiNameDeclaration
       List<ParameterDeclaration> parameters,
       boolean isClassInvocable,
       boolean isDefaultProperty,
-      Type type) {
+      Type type,
+      DelphiNameDeclaration readDeclaration,
+      DelphiNameDeclaration writeDeclaration) {
     super(location);
     this.parameters = parameters;
     this.isClassInvocable = isClassInvocable;
     this.isDefaultProperty = isDefaultProperty;
     this.type = type;
+    this.readDeclaration = readDeclaration;
+    this.writeDeclaration = writeDeclaration;
   }
 
   private static List<ParameterDeclaration> extractParameters(
@@ -62,10 +75,59 @@ public final class PropertyNameDeclaration extends AbstractDelphiNameDeclaration
     return node.getType();
   }
 
+  private static DelphiNameDeclaration extractReadDeclaration(
+      PropertyNode node, @Nullable PropertyNameDeclaration concreteDeclaration) {
+    if (concreteDeclaration != null) {
+      return concreteDeclaration.getReadDeclaration();
+    }
+
+    PropertyReadSpecifierNode readSpecifier = node.getReadSpecifier();
+    if (readSpecifier != null) {
+      return extractSpecifierDeclaration(readSpecifier.getExpression());
+    }
+
+    return null;
+  }
+
+  private static DelphiNameDeclaration extractWriteDeclaration(
+      PropertyNode node, @Nullable PropertyNameDeclaration concreteDeclaration) {
+    if (concreteDeclaration != null) {
+      return concreteDeclaration.getWriteDeclaration();
+    }
+
+    PropertyWriteSpecifierNode writeSpecifier = node.getWriteSpecifier();
+    if (writeSpecifier != null) {
+      return extractSpecifierDeclaration(writeSpecifier.getExpression());
+    }
+
+    return null;
+  }
+
+  private static DelphiNameDeclaration extractSpecifierDeclaration(PrimaryExpressionNode node) {
+    DelphiNameDeclaration result = null;
+    for (int i = 0; i < node.jjtGetNumChildren(); ++i) {
+      Node child = node.jjtGetChild(i);
+      if (child instanceof NameReferenceNode) {
+        result = ((NameReferenceNode) child).getLastName().getNameDeclaration();
+      }
+    }
+    return result;
+  }
+
   @Override
   @NotNull
   public Type getType() {
     return type;
+  }
+
+  @Nullable
+  public DelphiNameDeclaration getReadDeclaration() {
+    return readDeclaration;
+  }
+
+  @Nullable
+  public DelphiNameDeclaration getWriteDeclaration() {
+    return writeDeclaration;
   }
 
   @Override
@@ -105,7 +167,17 @@ public final class PropertyNameDeclaration extends AbstractDelphiNameDeclaration
             .collect(Collectors.toUnmodifiableList()),
         isClassInvocable,
         isDefaultProperty,
-        type.specialize(context));
+        type.specialize(context),
+        specializeIfNotNull(readDeclaration, context),
+        specializeIfNotNull(writeDeclaration, context));
+  }
+
+  private static DelphiNameDeclaration specializeIfNotNull(
+      DelphiNameDeclaration declaration, TypeSpecializationContext context) {
+    if (declaration != null) {
+      declaration.specialize(context);
+    }
+    return null;
   }
 
   @Override
@@ -115,7 +187,8 @@ public final class PropertyNameDeclaration extends AbstractDelphiNameDeclaration
     }
     PropertyNameDeclaration that = (PropertyNameDeclaration) other;
     return that.node.getImage().equalsIgnoreCase(node.getImage())
-        && parameters.equals(that.parameters);
+        && parameters.equals(that.parameters)
+        && type.is(that.type);
   }
 
   @Override