+++ src/test/resources/org/sonar/plugins/delphi/symbol/generics/PropertySpecialization.pas
@@ -1,0 +1,25 @@
+unit PropertySpecialization;
+
+{This is a sample Delphi file.}
+
+interface
+
+ type
+   TFoo<T> = class
+     FData: T;
+     property Data: T read FData;
+   end;
+
+implementation
+
+procedure ConsumeString(Arg: String);
+begin
+  // Do nothing
+end;
+
+procedure Test(Foo: TFoo<String>);
+begin
+  ConsumeString(Foo.Data);
+end;
+
+end.